+++ src/indexer/mavryk/handlers/on_lending_controller_set_collateral_token.py
@@ -28,7 +28,7 @@ async def on_lending_controller_set_collateral_token(
         collateral_token_staked                     = collateral_token_storage.isStakedToken
         collateral_token_staking_contract_address   = collateral_token_storage.stakingContractAddress
         collateral_token_total_deposited            = float(collateral_token_storage.totalDeposited)
-        collateral_token_max_deposited_amount       = float(collateral_token_storage.maxDepositAmount)
+        collateral_token_max_deposit_amount         = float(collateral_token_storage.maxDepositAmount)
         collateral_token_paused                     = collateral_token_storage.isPaused
         collateral_token_id                         = 0
         collateral_token_standard                   = ""
@@ -74,6 +74,6 @@ async def on_lending_controller_set_collateral_token(
         lending_controller_collateral_token.is_staked_token                             = collateral_token_staked
         lending_controller_collateral_token.staking_contract_address                    = collateral_token_staking_contract_address
         lending_controller_collateral_token.total_deposited                             = collateral_token_total_deposited
-        lending_controller_collateral_token.max_deposited_amount                        = collateral_token_max_deposited_amount
+        lending_controller_collateral_token.max_deposit_amount                          = collateral_token_max_deposit_amount
         lending_controller_collateral_token.paused                                      = collateral_token_paused
         await lending_controller_collateral_token.save()